Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/fetch/cross-origin-resource-policy/scheme-restriction.any-expected.txt

CONSOLE MESSAGE: Cancelled load to https://127.0.0.1:9443/fetch/cross-origin-resource-policy/resources/hello.py?corp=same-site because it violates the resource's Cross-Origin-Resource-Policy response header.
CONSOLE MESSAGE: Fetch API cannot load https://127.0.0.1:9443/fetch/cross-origin-resource-policy/resources/hello.py?corp=same-site due to access control checks.

PASS Cross-Origin-Resource-Policy: same-site blocks retrieving HTTPS from HTTP 
